refactor: 폴더명 컨벤션의 일관성을 위해 constants, utils 폴더명을 각각 단수형으로 변경
