+++ src/domain/Race.js
@@ -1,6 +1,6 @@
 import Car from './Car';
-import CONFIG from '../constants/config';
-import pickRandomNumber from '../utils/pickRandomNumber';
+import CONFIG from '../constant/config';
+import pickRandomNumber from '../util/pickRandomNumber';
 
 class Race {
   #carList;
